rename files for merge
